Backend/requirements.txt
# =========================
# Core Framework
# =========================
fastapi==0.110.0
uvicorn[standard]==0.27.1

# =========================
# Configuration & Environment
# =========================
python-dotenv==1.0.1
pydantic-settings==2.2.1

# =========================
# Database (psycopg v3)
# =========================
sqlalchemy==2.0.43
psycopg[binary]==3.3.2
alembic==1.13.1

# =========================
# Authentication & Security
# =========================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# =========================
# Observability
# =========================
prometheus-client==0.19.0

# =========================
# Testing
# =========================
pytest==8.0.0
httpx==0.26.0
